Imported AuthController into register.js and cleaned up register.js, registration now works on the /Register route
