--- Backend/routes/Auth.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-const AuthController = require ('../controllers/AuthController')
-
-router.post('/', AuthController.register)
-
-module.exports = router
