+++ Backend/server.js
@@ -9,7 +9,6 @@ const fuelQuote = require('./routes/FuelQuoteModule')
 const quoteTable = require('./routes/QuoteTableModule')
 const LoginModuleRoute = require("./routes/LoginModule")
 const registerRoute = require('./routes/Register')
-const AuthRoute = require('./routes/Auth')
 
 //connect to the database
 mongoose.connect(process.env.DATABASE_URI).catch(error => console.log(error));
@@ -25,7 +24,6 @@ app.use("/api/clientprofilemanagement", ClientProfileManagementRoute);
 app.use('/api', fuelQuote, quoteTable)
 app.use("/Login", LoginModuleRoute);
 app.use('/Register', registerRoute);
-app.use('/Register1', AuthRoute);
 
 mongoose.connection.once('open', () => {
   console.log('Connected to MongoDB');
